requirements.txt
pandas~=1.3.5
bs4~=0.0.1
beautifulsoup4~=4.10.0
selenium~=4.1.0
setuptools==60.5.0
python-kasa~=0.4.0
webdriver_manager
requests~=2.27.1
matplotlib~=3.5.1